LPS-74277 portal-search-web: Test - clear parameters from URL.
When one moves from a custom range to a labeled one (or vice versa), the
parameters of the previous ranges should be cleaned out.

Here we have tests to ensure this. First, we ensure
FacetUtil.removeURLParameters() will stay correct and also remove the
"from"/"to" parameters.

The parameters should also be removed from the back-end-generated URLs. So here
is a test to ensure labeled ranges URLs do not preserve custom range
parameters.

https://issues.liferay.com/browse/LPS-74277