LPS-74277 portal-search-web: Hide facet body on empty searches
Modified facet can always refine a non-empty search, even if only through the
custom range. However, if no value is returned, it obviously cannot do anything
and should not be rendered.

So, to know if modified facet can tune the search, we need to know the number
of hits from the query. If it is positive, then modified facet can make an
effect and thus should appear.

Now, if the number of hits is zero but the modified facet is already
restricting the query, then the modified facet should be rendered, because the
modified facet can be the reason the query is empty. The user should have an
option to reset the facet!

Finally, if there is no result and no range is selected in the modified facet,
the modified facet should not be displayed.

https://issues.liferay.com/browse/LPS-74277